- Fixed a double-fclose and memory leaks in zipdir. - Fixed building of LZMA SDK on Linux with some help from p7zip.
SVN r1469 (trunk)

+++ src/files.cpp
@@ -378,7 +378,7 @@ FileReaderLZMA::FileReaderLZMA (FileReader &file, size_t uncompressed_size, bool
 	OutProcessed = 0;
 
 	// Read zip LZMA properties header
-	if (File.Read(header, sizeof(header)) < sizeof(header))
+	if (File.Read(header, sizeof(header)) < (long)sizeof(header))
 	{
 		I_Error("FileReaderLZMA: File too shart\n");
 	}
